feat: 添加网关基础模型、存储层分页查询及消息处理功能
新增网关基础模型定义，包括分页结构、请求响应模型和常量定义
实现存储层各模块的分页查询功能，包括通道、绑定、MCP、参数、技能、工具、提供者、会话和消息
重构内存存储为消息存储，支持消息的分页查询和搜索
添加网关处理器，包括通用、聊天、通道、MCP、消息、会话、任务和提供者等模块的基础CRUD和分页查询功能
新增常量定义测试用例

+++ agent/icooclaw/pkg/storage/model.go
@@ -12,9 +12,9 @@ import (
 const tableNamePrefix = "icooclaw_"
 
 type Model struct {
-	ID        string    `gorm:"column:id;type:char(36);primaryKey;comment:主键UUID" json:"id"`           // 主键 uuid
-	CreatedAt time.Time `gorm:"column:created_at;type:datetime;comment:创建时间" json:"created_at"`       // 创建时间
-	UpdatedAt time.Time `gorm:"column:updated_at;type:datetime;comment:更新时间" json:"updated_at"`       // 更新时间
+	ID        string    `gorm:"column:id;type:char(36);primaryKey;comment:主键UUID" json:"id"`    // 主键 uuid
+	CreatedAt time.Time `gorm:"column:created_at;type:datetime;comment:创建时间" json:"created_at"` // 创建时间
+	UpdatedAt time.Time `gorm:"column:updated_at;type:datetime;comment:更新时间" json:"updated_at"` // 更新时间
 }
 
 func (c *Model) BeforeCreate(tx *gorm.DB) error {
@@ -58,7 +58,7 @@ func (s StringArray) String() string {
 }
 
 type Page struct {
-	Size  int `json:"size"`
-	Page  int `json:"page"`
-	Total int `json:"total"`
+	Size  int   `json:"size"`
+	Page  int   `json:"page"`
+	Total int64 `json:"total"`
 }
